Remove unused Rseg gem
Rseg is "a Chinese Word Segmentation(中文分词) routine in pure Ruby." It
was added in 68e61b1e242e239dcd95147d94c45e7427d8a254. It hasn't been
updated since 2010. It seems not to be used anywhere, unless it's an
implicit dependency of some other gem.

The specs pass without it.